azuredataprocessor/Azure_poc_controller/requirements.txt
azure-eventhub
azure-data-tables